Specify a Java version (8) to try to fix Travis build
Previously, the code was passing with Travis, but Travis was picking a
default of Java 8. Now it's picking a default of Java 11. And it is
failing in tests:

Exception in thread "main" java.lang.reflect.InvocationTargetException
Caused by: java.lang.NoSuchFieldException: $jacocoAccess
FATAL ERROR in native method: processing of -javaagent failed

As this code was originally written to work with pre-Java 8, at least
specify Java 8 to see if we can get back to a successful Travis build.