build: Java 8 version
Revert to building with Java-8-era JDK and POM.
This returns to the way the code was built before adding
the Java 6 build support. It involves "undoing" the more modern
Java 11 approach that we use as the default for Java 6 devs.

At some point this too might be modernized either as a newer way
of building Java 8 version, or for later versions.