test: tighten exception spec in tests of assertions
Although our general approach for testing exceptions has been to have a
specific test to verify the exact type of exception thrown, then use a
much more general type of exception in extra tests of the exception's
message, we don't want to apply in in these cases. First, we don't
generally want to catch Throwable, it's just _too_ general and could
include real problems we don't want to catch. Secondly, in these
specific tests we are not testing general exception-throwing code, we
are testing test assertions. So in this case we want to emphasize that,
and avoid any confusion, by being as specific as possible.