style: modernize test subject names
Prefer "subject" to "test", as test appears everywhere, and it's
confusing to call the test subject "test", as it makes it sound like
it's a test, not a test subject. Prefer "subject" to the less
recognized "sut" (Subject Under Test). The subject within a test should
be understood in that context - that is, it's the subject of the test.