fix: OtherTimeZone must differ by offset significantly
There was a possibility that a "different" time zone could be used that
was not different enough, for example GMT, UTC, or London (in GMT not
BST) are different time zones, but could have the same time offsets.
This means they would not trigger failures in time-zone-specific (or
rather, offset-specific) tests.

This was fixed by requiring a "significant" difference in time offset,
not just a different time zone. Details of what is "significant" is
given in the Javadoc and unit tests.