refactor(test): clarify time difference is 3 hours
We look for different time zones with at with least a 3 hours
difference. We also look for time zones that are not too close to 12
hours apart, as that can allow 12-hour clock time differences to not
trigger. The buffer is again 3 hours. So tweak the code to make it
clear that the "9 hours" is actually 12 hours minus the three hour
buffer.