fix(test): ZonedTester can always find enough zones
Test and fix ZonedTester so that the exclusiveZones() can always find
enough zones to test in.
The first fix is to improve the function that finds zones with a
12-hour difference: if it can't find a +12 hours zone, it looks for a
-12 hours zone. The second fix is that if it can't find either a +12
_or_ a -12, then it constructs a custom zone with the appropriate
distance. As an example, America/St_Johns had an offset of -3.5 hours:
there aren't many time zones with half hours, so not always possible to
find one exactly 12 hours from this.