refactor(test): move @AfterClass for clarity
The @AfterClass method had ended up in the middle of the tests, which
makes for poor visibility. Happened as a result of putting at the _end_
of the tests, and then adding new tests without noticing.
As this is an unusual check, put it at the top rather than the bottom;
we usually favour "hiding" a lot of tedious set-up at the bottom, so
the tests can be found and read easily, but in this case it is unusual
enough to deserve being at the top.