test: check some tests in other time zones
Begin using our OtherTimeZone @Rule to verify that some tests work in
time zones other than just the developer's time zone.
Note that this has not been added to all because some tests in fact
fail in other time zones, and it is yet to be established whether it's
the nature of the test itself that fails, or the code under test.